完善user和article，增加搜索

+++ src/main/java/org/tomjerry/sweethome/service/implement/ArticleServiceImpl.java
@@ -1,11 +1,16 @@
 package org.tomjerry.sweethome.service.implement;
 
 import org.springframework.stereotype.Service;
+import org.springframework.util.ReflectionUtils;
 import org.tomjerry.sweethome.pojo.entity.ArticleEntity;
+import org.tomjerry.sweethome.pojo.entity.UserEntity;
 import org.tomjerry.sweethome.repository.ArticleRepository;
 import org.tomjerry.sweethome.service.ArticleService;
 
+import java.lang.reflect.Field;
+import java.sql.Timestamp;
 import java.util.List;
+import java.util.Map;
 
 @Service
 public class ArticleServiceImpl implements ArticleService {
@@ -15,10 +20,7 @@ public class ArticleServiceImpl implements ArticleService {
         this.articleRepository = articleRepository;
     }
 
-    @Override
-    public void addArticle(ArticleEntity article) {
-        articleRepository.save(article);
-    }
+
 
     public ArticleEntity addArticle(String title, String content, int userId) {
         ArticleEntity article = new ArticleEntity();
@@ -29,45 +31,58 @@ public class ArticleServiceImpl implements ArticleService {
         return article;
     }
 
-    @Override
-    public void updateArticle(ArticleEntity article) {
-        articleRepository.save(article);
-    }
+
 
     @Override
-    public void deleteArticle(int id) {
-        articleRepository.deleteById(id);
+    public boolean updateArticle(Integer id, Map<String, Object> updates) {
+
+        ArticleEntity article = articleRepository.findById(id).orElse(null);
+        if (article == null) {
+            return false;
+        }
+
+        updates.forEach((key, value) -> {
+            Field field = ReflectionUtils.findField(UserEntity.class, key);
+            if (field == null) {
+                return;
+            }
+            field.setAccessible(true);
+            ReflectionUtils.setField(field, article, value);
+
+        });
+
+        //设置更新时间
+        article.setUpdate_time(new Timestamp(System.currentTimeMillis()));
+
+        articleRepository.save(article);
+
+        return true;
     }
+
+
+
+    @Override
+    public boolean deleteArticle(Integer id) {
+
+        if (!articleRepository.existsById(id)) {
+            return false;
+        }
+
+        articleRepository.deleteById(id);
+        return true;
+    }
+
+
 
     @Override
     public ArticleEntity getArticleById(int id) {
         return articleRepository.findById(id).orElse(null);
     }
 
-    @Override
-    public List<ArticleEntity> getArticlesByTitle(String title) {
-        return articleRepository.findByTitleContaining(title);
-    }
 
-    @Override
-    public List<ArticleEntity> getArticlesByContent(String content) {
-        return articleRepository.findByContentContaining(content);
-    }
-
-    @Override
-    public List<ArticleEntity> getArticlesByKeyword(String keyword) {
-        List<ArticleEntity> articleEntities = articleRepository.findByTitleContaining(keyword);
-        articleEntities.addAll(articleRepository.findByContentContaining(keyword));
-        return articleEntities;
-    }
 
     @Override
     public List<ArticleEntity> getArticlesByUserId(int userId) {
         return articleRepository.findByUserid(userId);
-    }
-
-    @Override
-    public List<ArticleEntity> getAllArticles() {
-        return articleRepository.findAll();
     }
 }
